news-aggregator/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
aiohttp
aiofiles
beautifulsoup4
feedparser
requests
tweepy
openai
anthropic
transformers
torch
spacy
pillow
pytesseract
openai-whisper
moviepy
pydub
python-docx
PyPDF2
mammoth
pandas
openpyxl
python-multipart
redis
celery
apscheduler
pydantic-settings
python-dotenv